Rename Survey to CurrentSurvey and CurrentSurvey to CurrentSurveyPortal.
See: https://github.com/google/site-kit-wp/issues/3380#issuecomment-853999963.
