+++ assets/js/components/surveys/CurrentSurvey.js
@@ -1,5 +1,5 @@
 /**
- * CurrentSurvey component.
+ * Survey component.
  *
  * Site Kit by Google, Copyright 2021 Google LLC
  *
@@ -17,31 +17,172 @@
  */
 
 /**
+ * WordPress dependencies
+ */
+import { useCallback, useEffect, useState } from '@wordpress/element';
+
+/**
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import { CORE_SITE } from '../../googlesitekit/datastore/site/constants';
+import { CORE_FORMS } from '../../googlesitekit/datastore/forms/constants';
 import { CORE_USER } from '../../googlesitekit/datastore/user/constants';
-import Survey from './Survey';
-import Portal from '../Portal';
-const { useSelect } = Data;
+import SurveyCompletion from './SurveyCompletion';
+import SurveyQuestionRating from './SurveyQuestionRating';
+const { useDispatch, useSelect } = Data;
 
 const CurrentSurvey = () => {
-	const usingProxy = useSelect( ( select ) => select( CORE_SITE ).isUsingProxy() );
-	const currentSurvey = useSelect( ( select ) => select( CORE_USER ).getCurrentSurvey() );
+	const [ hasSentSurveyShownEvent, setHasSentSurveyShownEvent ] = useState( false );
+	const [ hasSentCompletionEvent, setHasSentCompletionEvent ] = useState( false );
 
-	if ( ! usingProxy || ! currentSurvey ) {
+	useEffect( () => {
+		if ( questions?.length && ! hasSentSurveyShownEvent ) {
+			sendSurveyEvent( 'survey_shown' );
+			setHasSentSurveyShownEvent( true );
+		}
+	}, [ questions, hasSentSurveyShownEvent, sendSurveyEvent ] );
+
+	const completions = useSelect( ( select ) => select( CORE_USER ).getCurrentSurveyCompletions() );
+	const questions = useSelect( ( select ) => select( CORE_USER ).getCurrentSurveyQuestions() );
+	const surveySession = useSelect( ( select ) => select( CORE_USER ).getCurrentSurveySession() );
+
+	const formName = surveySession ? `survey-${ surveySession.session_id }` : null;
+
+	const answers = useSelect( ( select ) => select( CORE_FORMS ).getValue( formName, 'answers' ) );
+
+	const currentQuestionOrdinal = answers?.length ? answers.reduce( ( a, b ) => {
+		return ( a.question_ordinal > b.question_ordinal ? a.question_ordinal : b.question_ordinal );
+	} ).question_ordinal + 1 : 1;
+
+	const shouldHide = useSelect( ( select ) => select( CORE_FORMS ).getValue( formName, 'hideSurvey' ) );
+
+	const { setValues } = useDispatch( CORE_FORMS );
+	const { sendSurveyEvent } = useDispatch( CORE_USER );
+
+	const currentQuestion = questions?.find( ( question ) => {
+		return question.question_ordinal === currentQuestionOrdinal;
+	} );
+
+	const answerQuestion = useCallback( ( answer ) => {
+		// eslint-disable-next-line camelcase
+		if ( currentQuestion?.question_type === 'rating' ) {
+			sendSurveyEvent( 'question_answered', {
+				// eslint-disable-next-line camelcase
+				question_ordinal: currentQuestion?.question_ordinal,
+				answer: {
+					answer_ordinal: answer.answer_ordinal,
+				},
+			} );
+
+			setValues( formName, {
+				answers: [
+					...answers || [],
+					{
+					// eslint-disable-next-line camelcase
+						question_ordinal: currentQuestion?.question_ordinal,
+						answer_ordinal: answer.answer_ordinal,
+					},
+				],
+			} );
+		}
+	}, [ answers, currentQuestion, formName, sendSurveyEvent, setValues ] );
+
+	const ctaOnClick = useCallback( () => {
+		sendSurveyEvent( 'follow_up_link_clicked' );
+		sendSurveyEvent( 'survey_closed' );
+
+		setValues( formName, { hideSurvey: true } );
+	}, [ formName, sendSurveyEvent, setValues ] );
+
+	const dismissSurvey = useCallback( () => {
+		sendSurveyEvent( 'survey_closed' );
+
+		setValues( formName, { hideSurvey: true } );
+	}, [ formName, sendSurveyEvent, setValues ] );
+
+	// Check to see if a completion trigger has been met.
+	let triggeredCompletion;
+	// We only have trigger conditions for questions that are answered with
+	// ordinal values right now.
+	const ordinalAnswerMap = answers?.length ? answers.reduce( ( acc, answer ) => {
+		return {
+			...acc,
+			[ answer.question_ordinal ]: answer.answer_ordinal,
+		};
+	}, {} ) : [];
+
+	// Use Array.some to avoid looping through all completions; once a matching
+	// completion has been found, treat the survey as complete.
+	completions?.some( ( completion ) => {
+		completion.trigger_condition.some( ( condition ) => {
+			// If a question was answered with the appropriate value, a completion
+			// trigger has been fulfilled and we should treat this survey as complete.
+			if ( condition.answer_ordinal.includes( ordinalAnswerMap[ condition.question_ordinal ] ) ) { // eslint-disable-line camelcase
+				triggeredCompletion = completion;
+				return true;
+			}
+
+			return false;
+		} );
+
+		if ( triggeredCompletion ) {
+			return true;
+		}
+
+		return false;
+	} );
+
+	// If no specific trigger has been found but all questions are answered, use
+	// the first available trigger.
+	if ( ! triggeredCompletion && questions?.length && currentQuestionOrdinal > questions?.length ) {
+		triggeredCompletion = completions[ 0 ];
+	}
+
+	useEffect( () => {
+		if ( triggeredCompletion && ! hasSentCompletionEvent ) {
+			sendSurveyEvent( 'completion_shown', {
+				completion_ordinal: triggeredCompletion.completion_ordinal,
+			} );
+			setHasSentCompletionEvent( true );
+		}
+	}, [ hasSentCompletionEvent, sendSurveyEvent, triggeredCompletion ] );
+
+	if ( shouldHide || ! questions || ! completions ) {
 		return null;
 	}
 
-	return (
-		<Portal slug="survey">
-			<Survey
-				questions={ currentSurvey.question }
-				completions={ currentSurvey.completion }
+	if ( triggeredCompletion ) {
+		return (
+			<SurveyCompletion
+				dismissSurvey={ dismissSurvey }
+				ctaOnClick={ ctaOnClick }
+				ctaText={ triggeredCompletion.follow_up_text }
+				ctaURL={ triggeredCompletion.follow_up_url }
+				title={ triggeredCompletion.completion_title }
+			>
+				{ triggeredCompletion.completion_text }
+			</SurveyCompletion>
+		);
+	}
+
+	// eslint-disable-next-line camelcase
+	if ( currentQuestion?.question_type === 'rating' ) {
+		return (
+			<SurveyQuestionRating
+				answerQuestion={ answerQuestion }
+				choices={ currentQuestion.question.answer_choice }
+				dismissSurvey={ dismissSurvey }
+				key={ currentQuestion.question_ordinal }
+				question={ currentQuestion.question_text }
 			/>
-		</Portal>
-	);
+		);
+	}
+
+	return null;
 };
+
+CurrentSurvey.propTypes = {};
+
+CurrentSurvey.defaultProps = {};
 
 export default CurrentSurvey;
